move fetch gas estimates to utils
